merge

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -3,6 +3,9 @@ using AuctionService.DTOs;
 using AuctionService.Entities;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
+using Contracts;
+using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -14,11 +17,13 @@ public class AuctionsController : ControllerBase
 {
     private readonly AuctionDbContext _context;
     private readonly IMapper _mapper;
+    private readonly IPublishEndpoint _publishEndpoint;
 
-    public AuctionsController(AuctionDbContext context, IMapper mapper)
+    public AuctionsController(AuctionDbContext context, IMapper mapper, IPublishEndpoint publishEndpoint)
     {
         _context = context;
         _mapper = mapper;
+        _publishEndpoint = publishEndpoint;
     }
 
     [HttpGet]
@@ -51,12 +56,18 @@ public class AuctionsController : ControllerBase
 
     }
 
+    [Authorize]
     [HttpPost]
-    public async Task<ActionResult<AuctionDto>> CreateNewAuction(CreateAuctionDto auctionDto)
+    public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        auction.Seller = "test";
+        // auction.Seller = "test";
+        auction.Seller = User.Identity.Name;
         _context.Auctions.Add(auction);
+
+        var newAuction = _mapper.Map<AuctionDto>(auction);
+
+        await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
 
         var result = await _context.SaveChangesAsync() > 0;
 
@@ -64,8 +75,10 @@ public class AuctionsController : ControllerBase
 
         return CreatedAtAction(nameof(GetAuctionById),
                                 new { auction.Id },
-                                 _mapper.Map<AuctionDto>(auction));
+                                 newAuction);
     }
+
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -74,13 +87,19 @@ public class AuctionsController : ControllerBase
 
         if (auction is null) return NotFound();
 
-        // TODO: check seller == username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
         auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
         auction.Item.Mileage = updateAuctionDto.Mileage ?? auction.Item.Mileage;
         auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
+
+
+        var auctionUpdatedDto = _mapper.Map<AuctionDto>(auction);
+
+        // publish message update
+        await _publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auctionUpdatedDto));
 
         var result = await _context.SaveChangesAsync() > 0;
 
@@ -89,6 +108,8 @@ public class AuctionsController : ControllerBase
         return BadRequest("Problem saving changes");
 
     }
+
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -96,9 +117,11 @@ public class AuctionsController : ControllerBase
 
         if (auction is null) return NotFound();
 
-        //TODO: check seller == username
+        if(auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
+
+        await _publishEndpoint.Publish<AuctionDeleted>(new { Id = id.ToString() });
 
         var result = await _context.SaveChangesAsync() > 0;
